refactor: SessionTest 분리
