+++ src/main/java/nextstep/courses/domain/Session.java
@@ -1,6 +1,5 @@
 package nextstep.courses.domain;
 
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
